Email password authentication done

+++ firebase/index.js
@@ -15,6 +15,7 @@ import {
   deleteDoc,
   doc
 } from "firebase/firestore";
+import {getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, onAuthStateChanged, signOut} from "firebase/auth";
 
 // Your web app's Firebase configuration
 const firebaseConfig = {
@@ -29,5 +30,6 @@ const firebaseConfig = {
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
+const auth = getAuth(app);
 
-export { collection, addDoc, db, getDocs, query, where, onSnapshot, serverTimestamp, deleteDoc, doc};
+export { collection, addDoc, db, getDocs, query, where, onSnapshot, serverTimestamp, deleteDoc, doc, auth, createUserWithEmailAndPassword, signInWithEmailAndPassword, onAuthStateChanged, getAuth, signOut};
